% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{a}
\alias{a}
\title{a}
\description{
Example data a. An array with 3-D fluoresence  results. The 3 dimensions are the excitation wavelength (character),
the emission wavelength (character), and the sample number, e.g.,"GRnumber" (character).
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{data}
\keyword{data,}
\keyword{fluoresence}
\keyword{optical}
\keyword{vectorized}

