\docType{data}
\name{dfsags}
\alias{dfsags}
\title{dfsags}
\description{
  Example data dfsags. This dataframe contains three
  columns. The first two columns contain the lower and
  upper wavelength in nm (as integer) for which a spectral
  slope is to be calculated for each sample. The third
  column "Name" contains the name of the spectral slope
  which can then be used as a summary optical variable.
}
\author{
  Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{absorption,}
\keyword{data}
\keyword{datasets}
\keyword{optical}
\keyword{slopes,}
\keyword{spectral}

