\docType{data}
\name{ex_ems}
\alias{ex_ems}
\title{ex_ems}
\description{
  Example data ex_ems. This dataframe contains three
  columns. The first column called "Peak" contains the
  character name of the EEMs peak. The second column called
  "ExCA" contains the excitation wavelength for a
  particular peak as type integer. The third column called
  "EmCA" contains the emission wavelength for a particular
  peak as type integer.
}
\author{
  Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{EEMs,}
\keyword{absorption,}
\keyword{data}
\keyword{datasets}
\keyword{excitation-emission,}
\keyword{fluoresence,}
\keyword{optical}

