% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\docType{data}
\name{fractionSignals}
\alias{fractionSignals}
\title{fractionSignals}
\description{
Example data fractionSignals. Contains one column "variable1" with the name of the first optical metrics used as
and a second variable "variable2" to compute the fraction variable1/(variable1+variable2)
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{optical}
\keyword{properties}
