% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAbs}
\alias{getAbs}
\title{getAbs}
\usage{
getAbs(dataAbs, waveCol = "wavelength", signals, colSubsetString = "gr",
  dataSummary, grnum = "GRnumber")
}
\arguments{
\item{dataAbs}{dataframe with absorbance spectra results, one column per sample}

\item{waveCol}{Column name to define the wavelengths for which absorbance was measured}

\item{signals}{vector with absorbance wavelengths to extract.}

\item{colSubsetString}{unique characters to identify which columns have absorbance data.
The default is "gr" to comply with the common naming from the CA WSC}

\item{dataSummary}{dataframe with summary absorbance and fluoresence data. This
function adds columns to the end of this dataframe as additional summary data.}

\item{grnum}{Column name that defines the grnumbers in the dataSummary dataframe.
These names are used to merge spectral slope data into the summary dataframe.
This function assumes the column names of the dataAbs are grnumbers as well.}
}
\description{
Retrieves individual absorbance coefficients
}
\examples{
dataAbs <- dfabs
waveCol <- "wavelength"
dfSum <- data.frame(grnumber=grep("gr",names(dfabs),value=TRUE))
dataSummary <- dfsummary
signals <- signals
colSubsetString <- "gr"
grnum <- "GRnumber"
test <- getAbs(dataAbs=dfabs,waveCol="wavelength",signals=signals,
              colSubsetString="gr",dataSummary=dataSummary,grnum="GRnumber")
}

