% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExpResid}
\alias{getExpResid}
\title{getSag}
\usage{
getExpResid(wavelngth, rangeReg, rangeGap, dataAbs, waveCol = "wavelength",
  colSubsetString = "gr", dataSummary, grnum = "GRnumber")
}
\arguments{
\item{dataAbs}{dataframe with absorbance spectra results, one column per sample}

\item{waveCol}{Column name to define the wavelengths for which absorbance was measured}

\item{signals}{dataframe with three columns. The first column represents the low wavelength,
the second column represents the high wavelength for which spectral slopes are
to be defined, and the third column is the variable name to be used.
A spectral slope is computed for each row.}

\item{colSubsetString}{unique characters to identify which columns have absorbance data.
The default is "gr" to comply with the common naming from the CA WSC}

\item{dataSummary}{dataframe with summary absorbance and fluoresence data. This
function adds columns to the end of this dataframe as additional summary data.}

\item{grnum}{Column name that defines the grnumbers in the dataSummary dataframe.
These names are used to merge spectral slope data into the summary dataframe.
This function assumes the column names of the dataAbs are grnumbers as well.}
}
\description{
Computes spectral slopes from absorbance data using a linear regression to
determine the first order decay function as defined in Helms et al. 2008,
Limnol. Oceanogr., 53(3), 955-969. aL = aRef * exp(-S*(L-LRef)) where a = absorbance
coefficient, S = spectral slope, and L = wavelength.
}

