% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractionSumFxn.R
\name{getFrcSum}
\alias{getFrcSum}
\title{getFrcSum}
\usage{
getFrcSum(dataSummary, sigs, grnum)
}
\arguments{
\item{dataSummary}{dataframe with summary absorbance and fluoresence data.}

\item{sigs}{signals vector of variable names in dataSummary for generating fractions}

\item{grnum}{character column name that defines the column with sample names in the dataSummary dataframe.
These names are used to merge fraction data into the summary dataframe.}
}
\value{
dataSummary dataframe with the additional columns of spectral fractions computed using getFrcSum
}
\description{
Computes the fraction of the sum of two signals from optical data as:
Fraction = variable 1 / (variable 1 + variable 2)
}
\examples{
dataSummary <- dfsummary
sigs <- fractionSignals
grnum <- "GRnumber"
test <- getFrcSum(dataSummary,sigs,grnum)
}
