% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTransFxn.R
\name{getLog10}
\alias{getLog10}
\title{getLog10}
\usage{
getLog10(dataSummary, signals, grnum)
}
\arguments{
\item{dataSummary}{dataframe with summary absorbance and fluoresence data.}

\item{signals}{character vector of variable names in dataSummary for generating log transforms}

\item{grnum}{character column name that defines the column with sample names in the dataSummary dataframe.
These names are used to merge ratio data into the summary dataframe.}
}
\value{
dataframe with the log 10 transform of the summary absorbance and fluoresence data.
}
\description{
Computes log transform of optical summary data.
}
\examples{
dataSummary <- dfsummary
signals <- ratioSignals[which(ratioSignals[2]>0),1]
grnum<-"GRnumber"
test2 <- getLog10(dataSummary,signals,grnum)
}

