% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRatiosBase.R
\name{getRatiosBase}
\alias{getRatiosBase}
\title{getRatiosBase}
\usage{
getRatiosBase(dataSummary, sigs, baseSig, grnum, specifyOrder = FALSE,
  recordOrder = FALSE, ratioVars = FALSE)
}
\arguments{
\item{dataSummary}{dataframe with summary absorbance and fluoresence data.}

\item{sigs}{signals vector of variable names in dataSummary for generating ratios}

\item{baseSig}{name of the column to subtract out the base signal}

\item{grnum}{character column name that defines the column with sample names in the dataSummary dataframe.
These names are used to merge spectral slope data into the summary dataframe.}

\item{specifyOrder}{boolean variable to signify whether the numerator and denominator 
shold stay as specified in ratioVars}

\item{recordOrder}{boolean variable to signify whether the numerator and denominator
order should be recorded and used as output rather than the actual ratios.}

\item{ratioVars}{dataframe with specific ratios specified. Numerators in the first column
and denominators in the second column}
}
\value{
dataSummary dataframe with the additional columns of spectral ratios computed using getRatios
}
\description{
Computes ratios from optical data. Assumes that the signal with the greatest 
mean is in the numerator making the mean ratio greater than one.
}
\examples{
dataSummary <- dfsummary
sigs <- ratioSignals[which(ratioSignals[2]>0),1]
grnum <- "GRnumber"
baseSig <- "F"
test <- getRatiosBase(dataSummary,sigs = sigs, baseSig=baseSig, grnum = grnum)
}
