% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotAbsFxn.R
\name{plotAbs}
\alias{plotAbs}
\title{plotAbs}
\usage{
plotAbs(dataAbs, WaveCol, absCol, ...)
}
\arguments{
\item{dataAbs}{dataframe of absorbance data}

\item{WaveCol}{character column name to define the wavelengths for which absorbance was measured.}

\item{absCol}{character column name to define which column to plot}

\item{...}{additional plotting parameters as needed. xlim and ylim are commonly used here.}
}
\value{
Absorbance coefficient curve
}
\description{
Plot absorbance coefficient curves
}
\examples{
WaveCol <- "wavelengths"
absCol <- "gr13307"
titleSize <- 1.1
mainTitle <- "Example Absorbance Plot"
xlim <- c(239,450)
dataAbs <- dfabs
plotAbs(dataAbs,WaveCol,absCol,main=mainTitle,cex.main=titleSize,xlim=xlim)
}

