% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotEEMs2}
\alias{plotEEMs2}
\title{plotEEMs2}
\usage{
plotEEMs2(mat, Ex, Em, nlevels = 50, Peaks, peakCol = "Peak",
  peakEx = "ExCA", peakEm = "EmCA", mainTitle)
}
\arguments{
\item{mat}{2-D matrix of excitation-emmission spectra}

\item{Ex}{Excitation wavelengths}

\item{Em}{Emmission wavelengths}

\item{nlevels}{color levels for contour graph. 50 is commonly used for a value here.}

\item{Peaks}{dataframe with peaks to be indicated on the graph}

\item{peakEx}{column in Peaks to use for excitation wavelengths}

\item{peakEm}{column in Peaks to use for emmission wavelengths}

\item{mainTitle}{Plot title}
}
\description{
Plot contour graph of excitation emmission spectra with defined peaks indicated
on the graph
}

