% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataInfo.R
\docType{data}
\name{ratioSignals}
\alias{ratioSignals}
\title{ratioSignals}
\description{
Example data ratioSignals. Contains one column "ratioSignals" with the name of the different optical metrics used as
signals for different chemical species in freshwater.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{absorption}

