% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{signals}
\alias{signals}
\title{signals}
\description{
Example data signals. Contains one column called "Peak" with different excitation-emission peaks that act
as signals for particular chemical species. These peaks are well characterized and the "Source" column
in this dataframe lists the source that characterized each excitation-emission peak and the chemical species
that it identifies. There are four additional columns with the Excitation and Emission wavelengths for each peak.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{absorption}

