% Generated by roxygen2 (4.0.1): do not edit by hand
\name{VectorizedTo3DArray}
\alias{VectorizedTo3DArray}
\title{VectorizedTo3DArray}
\usage{
VectorizedTo3DArray(df, ExEm = "Wavelength.Pairs")
}
\arguments{
\item{df}{dataframe with vectorized fluorescence data in the format from the
CA WSC with the ExEm column along with one column per sample with GRnumber as
the header}

\item{ExEm}{Character name of column with Excitation and Emmission wavelengths
defined in this format: ###/###. For example Excitation 250 and emmission 400
would be represented as 250/400.}
}
\description{
Converts vectorized fluorescence dataframe into a 3-D array with Ex, Em, and
GRnumber as the dimensions. This results in one 2-D excitation-emmission array
per sample.
}

