% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\docType{data}
\name{dfabs}
\alias{dfabs}
\title{dfabs}
\description{
Example data dfabs. This dataframe contains the absorbance spectra for 265 samples. The wavelength
is measured from 200nm to 750nm for each sample. One column called "wavelengths" contains the wavelength
for the absorbance measurment in nm.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{absorption,}
\keyword{data}
\keyword{optical}
\keyword{wavelength,}
