% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataInfo.R
\docType{data}
\name{dfFluor}
\alias{dfFluor}
\title{dfFluor}
\description{
Example data dfFluor. Contains vectorized fluoresence data and one column called "Wavelength.Pairs" which
contains the Excitation and Emission wavelengths defined in this format: ###/###.
For example, Excitation 250 and emmission 400 would be represented as "250/400".
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{(EEM).}
\keyword{data,}
\keyword{data,excitation,}
\keyword{emission,}
\keyword{excitation-emission}
\keyword{fluoresence}
\keyword{fluoresence,}
\keyword{matrix}
\keyword{optical}
\keyword{vectorized}

