\docType{data}
\name{dfsummary}
\alias{dfsummary}
\title{dfsummary}
\description{
  Example data dfsummary. Contains summary optical
  variables for each sample, here called "GRnumber". These
  summary optical variables are computed using the
  functions in this package.
}
\author{
  Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{absorbance}
\keyword{absorption,}
\keyword{data.}
\keyword{datasets}
\keyword{emission,}
\keyword{excitation,}
\keyword{fluoresence,}
\keyword{freshness}
\keyword{humic}
\keyword{index,}
\keyword{optical}
\keyword{peaks,}
\keyword{ratios,}
\keyword{slopes,}
\keyword{spectral}

