% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExpResid}
\alias{getExpResid}
\title{getSag}
\usage{
getExpResid(wavelength, rangeReg, rangeGap, dataAbs, waveCol, colSubsetString,
  dataSummary, grnum)
}
\arguments{
\item{wavelength}{absorbance numeric wavelength}

\item{rangeReg}{numeric string with absorbance wavelength range to be considered for computing spectral slope}

\item{rangeGap}{numeric string with the absorbance wavelength range for which decay function should be applied}

\item{dataAbs}{dataframe with absorbance spectra results, one column per sample}

\item{waveCol}{character column name to define the wavelengths for which absorbance was measured}

\item{colSubsetString}{character with unique letters to identify which columns have absorbance data.
The default is "gr" to comply with the common naming from the CA WSC}

\item{dataSummary}{dataframe with summary absorbance and fluoresence data. This
function adds columns to the end of this dataframe as additional summary data.}

\item{grnum}{character column name that defines the grnumbers in the dataSummary dataframe.}
}
\value{
dataframe with the added spectral slope for each sample and plots with the absorbance spectra for rangeReg showing the
model constructed using the spectral slope (red); and the absorbance data where black = the data in rangeReg that is not in rangeGap and blue = the data from rangeGap.
}
\description{
Computes spectral slopes from absorbance data using a linear regression to determine the first order decay function as defined in Helms et al. 2008, Limnol. Oceanogr., 53(3), 955-969. aL = aRef * exp(-S*(L-LRef)) where a = absorbance coefficient, S = spectral slope, and L = wavelength. Function assumes that the column names of the absorbance data file being used are formatted as grnumbers.
}
\examples{
wavelength <- 267
rangeReg <- c(240,340)
rangeGap <- c(255,300)
dataAbs <- dfabs
waveCol <- "wavelengths"
colSubsetString <- "gr"
dataSummary <- dfsummary
grnum <- "GRnumber"
testdfOpt <- getExpResid(wavelength,rangeReg,rangeGap,dataAbs,waveCol,
                    colSubsetString,dataSummary,grnum)
}

