% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLog10}
\alias{getLog10}
\title{getLog10}
\usage{
getLog10(dataSummary, signals, grnum = "GRnumber")
}
\arguments{
\item{dataSummary}{dataframe with summary absorbance and fluoresence data.}

\item{signals}{vector of variable names in dataSummary for generating log transforms}

\item{grnum}{Column name that defines the grnumbers in the dataSummary dataframe.
These names are used to merge ratio data into the summary dataframe.}
}
\description{
Computes log transform of optical summary data.
}

