% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMeanFl}
\alias{getMeanFl}
\title{getMeanFl}
\usage{
getMeanFl(a, signals, Peak = "Peak", Ex1 = "Ex1", Ex2 = "Ex2",
  Em1 = "Em1", Em2 = "Em2", dataSummary, grnum = "GRnumber")
}
\arguments{
\item{a}{array with 3-D fluorescence  results,}

\item{signals}{dataframe defining the max and min excitation and the max and
min emmission wavelengths for which to compute averates.}

\item{Peak}{the column name for the column with parameters to be computed}

\item{Ex1}{The first excitation wavelength in the range}

\item{Ex2}{the second excitation wavelength in the range. This can be blank
if one specific excitation wavelength is used.}

\item{Em1}{the first emmission wavelength in the range}

\item{Em2}{the second emmission wavelength in the range. This can be blank
if one specific emmission wavelength is used.}

\item{dataSummary}{dataframe with summary absorbance and fluoresence data. This
function adds columns to the end of this dataframe as additional summary data.}

\item{grnum}{Column name that defines the grnumbers in the dataSummary dataframe.
These names are used to merge Fluorescence data into the summary dataframe.
This function assumes names of the third dimension in the 3-D array a are
grnumbers as well.}
}
\description{
and freshness index as defined by PARLANTI, E., K. WORZ, L. GEOFFROY, AND M. LAMOTTE.
2000. Dissolved organic matter fluorescence spectroscopy as a tool to estimate
biological activity in a coastal zone submitted to anthropogenic inputs. Org.
Geochem. 31: 1765-1781, doi:10.1016/S0146-6380(00)00124-8
FreshI = ex310em380/max(ex310 between em470 and em520)
}
\examples{
Ex1 <- "Ex1"
Ex2 <- "Ex2"
Em1 <- "Em1"
Em2 <- "Em2"
grnum <- "GRnumber"
Peak <- "Peak"
dataSummary=dfsummary
testMeanFl <- getMeanFl(a,signals,dataSummary,grnum="GRnumber",Peak="Peak",Ex1="Ex1",
                        Ex2="Ex2",Em1="Em1",Em2="Em2")
}

